Changing the engine to use 16bpp mode (because that'll be needed for the Sega Saturn version in the end). This has the following consequences: - Movie playback currently disabled (ScummVM's movie API fails) - True shading in the conversation and inventory boxes - No need to adept those boxes to ever changing palettes - Palette errors related to Mike resolved - Palette merging errors eliminated (since there's no palette merging anymore)

+++ engines/darkseed2/resources.h
@@ -157,7 +157,7 @@ private:
 	ImageType _imageType; ///< The type of images the game uses.
 
 	uint16 _archiveCount; ///< Number of indexed archive files.
-	uint16 _resCount;  ///< Number of indexed resources.
+	uint32 _resCount;  ///< Number of indexed resources.
 
 	/** All indexed archives. */
 	Common::Array<Archive> _archives;
